[UPDATE]: mid third deliverable
